+++ appveyor.yml
@@ -18,7 +18,7 @@ install:
 
 build_script:
   - npm run lint
-  - npm run package -- --all
+  - npm run build
 
 before_test:
   - git config --global user.email "test@testy.com"
